Fixups for running pytest via VSCode (#933)
# Pull Request

## Title

Fixups for running pytest via VSCode

______________________________________________________________________

## Description

@motus noticed that running all tests in VSCode was hanging inside a
DevContainer on MacOS.

After some debugging I determined that there was some mismatch between
discovering tests and running them, specifically for the
GridSearchOptimizer. This was true on the CLI as well.

Adding it to the `__init__.py` for that submodule seems to fix the
issue.

______________________________________________________________________

## Type of Change

- 🛠️ Bug fix

______________________________________________________________________

## Testing

- Local VSCode in both DevContainer and MacOS native mode.

______________________________________________________________________

## Additional Notes (optional)

Don't know of others like this but I wonder if that's an issue now ...

______________________________________________________________________