Pyright Improvements (#932)
# Pull Request

## Title

Pyright improvements

______________________________________________________________________

## Description

Pyright is a type checker that ships with VSCode's Pylance by default.

It is billed as a faster, though less complete, version of mypy.

As such it gets a few things a little differently that mypy and alerts
in VSCode.

This PR fixes those ("standard") alerts and removes the mypy extension
from VSCode's default extensions for MLOS in favor of just using pyright
(there's no sense in running both interactively). We do not enable
pyright's "strict" mode.

Additionally, it enables pyright in pre-commit rules to ensure those
fixes remain.

We leave the rest of the mypy checks as well since they are still
useful.

A list of some of the types of fixes:

- TypeDict initialization checks for Tunables
- Check that json.loads() returns a dict and not a list (e.g.)
- Replace ConcreteOptimizer TypeVar with a TypeAlias
- Add BoundMethod protocol for checking __self__ attribute
- Ensure correct type inference in a number of places
- Add `...` to Protocol methods to make pyright aware of the lack of
method body.
- Fix a few type annotations

______________________________________________________________________

## Type of Change

- 🛠️ Bug fix
- 🔄 Refactor
______________________________________________________________________

## Testing

- Additional CI checks as described above.

______________________________________________________________________

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>