Fixing an args type bug in launcher tunable_values processing (#958)
# Pull Request

## Title

Fixing an args type bug in launcher tunable_values processing

______________________________________________________________________

## Description

If tunable_values are specified in both a config file and on the cli
args, but the config file specifies them as a string, python will throw
an exception when trying to concat a string and a list. The json schema
allows either a list of paths or a single path, so this is valid in that
sense.

This also fixes the processing function args to only accept a list in
both cases.

______________________________________________________________________

## Type of Change

- 🛠️ Bug fix
______________________________________________________________________

## Testing

Locally

______________________________________________________________________