bugfix: local env not passed to the child process when shell_env_params present (#959)
Copy local environment variables and explicitly pass them to the child
process.

## Description

When we explicitly provide the environment variables to pass to the
`subprocess.run()` call, only those variables get passed to the child
process. That prevents the child process from seeing the environment
variables of the parent MLOS process. We've fixed that issue for MLOS on
Windows but not for other systems. This PR fixes it for Linux and MacOS.

## Testing

Update the existing unit tests to check for MLOS parameters *and* parent
environment variables being passed to the child process.