Delay entering TrialRunner context until run_trial (#970)
# Pull Request

## Title

Delay entering `TrialRunner` context until `run_trial`.

______________________________________________________________________

## Description

This is part of an attempt to try and see if can work around issues with
`multiprocessing.Pool` needing to pickle certain objects when forking.

For instance, if the Environment is using an SshServer, we need to start
an EventLoopContext in the background to handle the SSH connections and
threads are not picklable.

Nor are file handles, DB connections, etc., so there may be other things
we also need to adjust to make this work.

See Also #967

______________________________________________________________________

## Type of Change


- 🛠️ Bug fix
- 🔄 Refactor

______________________________________________________________________

## Testing

- Light so far (still in draft mode)
- Just basic existing CI tests (seems to not break anything)

______________________________________________________________________

## Additional Notes (optional)

I think this is incomplete. To support forking inside the Scheduler and
*then* entering the context of the given TrialRunner, we may also need
to do something about the Scheduler's Storage object.

That was true, those PRs are now forthcoming.  See Also #971 

For now this is a draft PR to allow @jsfreischuetz and I to play with
alternative organizations of #967.

______________________________________________________________________