Rename and reorganize some Scheduler methods (#975)
# Pull Request

## Title

Rename and reorganize some Scheduler methods

______________________________________________________________________

## Description

In preparation for ParallelScheduler PR (#971) this PR

1. Renames a few methods for clarity (e.g., `schedule_trial` -->
`add_trial_to_queue`) so that the role of other methods (e.g.,
`assign_trial_runners`) is more clear and readily overridable.
2. Moves some methods to the base class.
This is done based on an observation that some of the methods (e.g., the
main `start` loop) are largely reusable across both SyncScheduler and
ParallelScheduler if the method separation is done a little bit more
fine grained such that each can override the only the parts they need
to.

Additionall, it also

1. Also adds a subtle tweak on the `pending_trials` method in order to
filter based on whether or not the trial already has a runner assigned
or not.
2. Adjusts `TrialRunner.run_trial` to return the results of the Trial.
This isn't really used anywhere, but it is helpful to check that
TrialRunners run in a child process finished successfully by using that
result as a check in the return value, even though for Optimizer
bulk_registering it actually pulls the results back off of Storage in
the MainProcess.

______________________________________________________________________

## Type of Change

- 🔄 Refactor
- 🧪 Tests

______________________________________________________________________

## Testing

- [x] New CI tests for tweaks in `pending_trials`
- Existing CI tests

______________________________________________________________________

## Additional Notes (optional)


______________________________________________________________________

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>