Capture pytest conftest warnings (#1008)
# Pull Request

## Title

Capture pytest conftest warnings

______________________________________________________________________

## Description

Followup to #1007.

Copilot suggestions were correct. The `logger.warning()` messages
generated in `__init__.py` and `conftest.py` are captured, but not
emitted by default when *per-test* captures are done.

This switches those calls to `UserWarnings` so they are always captured.

This was the only location I could find for those.

Additionally, I adjust some checks for the `docker.sock` on Windows to
avoid some warnings.

______________________________________________________________________

## Type of Change

- 🛠️ Bug fix
- 🧪 Tests

______________________________________________________________________

## Testing

Manual.

______________________________________________________________________

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>