Bugfix: Fix crashing in postgres-backed storage (#1000)
# Pull Request

## Title

In PostgreSQL, attempting to insert a record with conflicting keys fails
(a benign action in MLOS), fails the entire transaction thus preventing
the update of other tables. Splitting the update into two independent
transactions solves the problem.
______________________________________________________________________

## Description

Update tables `trial_status` and `trial` in separate transactions
(that's OK for MLOS) to make sure PG won't fail the update of `trial` if
there is a duplicate record in `trial_status`.

- **Issue link**: Closes #999 

______________________________________________________________________

## Type of Change

- 🛠️ Bug fix

______________________________________________________________________

## Testing

Run TQP benchmark with PostgreSQL storage.